requirements.txt
playwright==1.51.0
python-pptx==0.6.23
Pillow==10.2.0
matplotlib==3.8.3 